refactor(WIP!): clean up main file
moved most of the logic into the routes dir.

another big thing is that there's a Router utility that allows us to set
the Post / Get / Put / Delete methods with ease, similar to Express.
