+++ backend/pw/check.go
@@ -1,0 +1,15 @@
+package pw
+
+import (
+	"golang.org/x/crypto/bcrypt"
+)
+// Hash user password from body
+func HashPassword(password string) (string, error) {
+	bytes, err := bcrypt.GenerateFromPassword([]byte(password), 5)
+	return string(bytes), err
+}
+
+func CheckPasswordHash(password, hash string) bool {
+	err := bcrypt.CompareHashAndPassword([]byte(hash), []byte(password))
+	return err == nil
+}
