+++ backend/utils/ctx.go
@@ -1,0 +1,3 @@
+package utils
+
+const CTX_KEY_DB = "db"
